requirements/base.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.5.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2025.6
cron_descriptor==2.0.8
cryptography==46.0.5
distro==1.9.0
Django==5.1.11
django-allauth==65.9.0
django-anymail==13.0
django-appconf==1.2.0
django-celery-beat==2.8.1
django-compressor==4.5.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==6.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==24.3
django-model-utils==5.0.0
django-redis==6.0.0
django-simple-history==3.7.0
django-timezone-field==7.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
fido2==2.1.1
flower==2.0.1
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
idna==3.11
inflection==0.5.1
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.5.4
openai==1.58.1
openpyxl==3.1.5
packaging==26.0
pillow==11.2.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg==3.2.9
psycopg-c==3.2.9
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==8.2
rcssmin==1.1.2
redis==6.2.0
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rjsmin==1.2.2
rpds-py==0.30.0
sentry-sdk==2.32.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
text-unidecode==1.3
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.9.0
